updated dependencies/config/imports/exports

+++ src/simulation/simulation-fork.ts
@@ -1,6 +1,7 @@
+import { BodyDataWithId } from './body'
 import { DEFAULT_MAX_MB } from './constants'
 
-import { BodyJson, Simulation, SimulationJson } from './simulation'
+import { Simulation, SimulationData } from './simulation'
 import { SimulationPhysical } from './simulation-physical'
 
 /* eslint-disable @typescript-eslint/no-var-requires */
@@ -18,8 +19,7 @@ const IS_CHILD = IS_NODE
 /**
  * Forks the physics integration into a separate child process
  */
-abstract class SimulationFork<B extends BodyJson> extends Simulation<B> {
-
+abstract class SimulationFork<B extends BodyDataWithId> extends Simulation<B> {
     private _childProcess: any = null
 
     public override get isRunning(): boolean {
@@ -29,31 +29,35 @@ abstract class SimulationFork<B extends BodyJson> extends Simulation<B> {
     // Override
 
     public override run() {
-
-        if (this.isRunning)
-            this.stop()
+        if (this.isRunning) this.stop()
 
         if (IS_BROWSER) {
-
-            const { default: WebWorker } = require('worker-loader!' + __filename)
+            const { default: WebWorker } = require('worker-loader!' +
+                __filename)
 
             this._childProcess = new WebWorker()
-            this._childProcess.onmessage = (msg: MessageEvent<SimulationJson['bodies']>) => this._update(msg.data)
+            this._childProcess.onmessage = (
+                msg: MessageEvent<SimulationData['bodies']>
+            ) => this._update(msg.data)
             this._childProcess.postMessage(this.toJSON())
         }
 
         if (IS_NODE) {
-
             const { fork } = require('child_process')
 
             const FORK_PATH = __filename
                 .replace('/src/', '/lib/')
                 .replace(/\.ts$/, '.js')
 
-            const FORK_MEMORY = { execArgv: [`--max-old-space-size=${DEFAULT_MAX_MB}`] }
+            const FORK_MEMORY = {
+                execArgv: [`--max-old-space-size=${DEFAULT_MAX_MB}`]
+            }
 
             this._childProcess = fork(FORK_PATH, FORK_MEMORY)
-            this._childProcess.on('message', (bodies: SimulationJson['bodies']) => this._update(bodies))
+            this._childProcess.on(
+                'message',
+                (bodies: SimulationData['bodies']) => this._update(bodies)
+            )
             this._childProcess.send(this.toJSON())
         }
 
@@ -61,50 +65,36 @@ abstract class SimulationFork<B extends BodyJson> extends Simulation<B> {
     }
 
     public override stop(): void {
+        if (IS_BROWSER) this._childProcess?.terminate()
 
-        if (IS_BROWSER)
-            this._childProcess?.terminate()
-
-        if (IS_NODE)
-            this._childProcess?.kill()
+        if (IS_NODE) this._childProcess?.kill()
 
         this._childProcess = null
     }
-
 }
 
 //// Execute ////
 
 if (IS_CHILD) {
-
-    const runSimulation = (input: SimulationJson, onTick: (output: SimulationJson['bodies']) => void) => {
+    const runSimulation = (
+        input: SimulationData,
+        onTick: (output: SimulationData['bodies']) => void
+    ) => {
         const simulation = new SimulationPhysical(input)
         simulation.on('tick', onTick)
         simulation.run()
     }
 
     if (IS_BROWSER)
-        self.onmessage = ({ data: input }: MessageEvent<SimulationJson>) =>
-            runSimulation(
-                input,
-                output => self.postMessage(output)
-            )
+        self.onmessage = ({ data: input }: MessageEvent<SimulationData>) =>
+            runSimulation(input, output => self.postMessage(output))
 
     if (IS_NODE)
-        process.on(
-            'message',
-            input => runSimulation(
-                input,
-                output => process.send?.(output)
-            )
+        process.on('message', input =>
+            runSimulation(input, output => process.send?.(output))
         )
-
 }
 
 //// Exports ////
 
-export default SimulationFork
-
-export {
-    SimulationFork
-}
+export { SimulationFork }
